// src/medical-history-record/interfaces/medical-history.interface.ts
export interface IMedicalHistoryRecord {
    previousIllnesses: string;
    consultationReason: string;
    respiratoryRate: number;
    heartRate: number;
    pulse: string;
    CRT: number;
    temperature: number;
    limphaticNodes: string;
    mucosa: string;
    temperament: string;
    findings: string;
    tests?: string;
}